--- products/business_knowledge/backend/facade/__init__.py
@@ -1,1 +1,0 @@
-
